refactor: rename typedMessages to typed-messages
Amp-Thread-ID: https://ampcode.com/threads/T-019d9e31-81dc-77a0-a9c4-1af7c1c36397
